# tools/Typhon.Workbench/ClientApp/schema/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Typhon.Workbench | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5200/"
    }
  ],
  "paths": {
...
    "/api/sessions/{sessionId}/topology": {
      "get": {
        "tags": [
          "Data"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopologyDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/queries/who-writes/{component}": {
      "get": {
        "tags": [
          "Data"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "component",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/queries/who-reads/{component}": {
      "get": {
        "tags": [
          "Data"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "component",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemListDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/tracks": {
      "get": {
        "tags": [
          "Data"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracksResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/track/{trackId}": {
      "get": {
        "tags": [
          "Data"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "trackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "uint32",
              "default": 0
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "uint32",
              "default": 4294967295
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackDataResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/aggregate": {
      "post": {
        "tags": [
          "Data"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AggregationRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AggregationRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AggregationRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AggregationResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}/subscribe": {
      "post": {
        "tags": [
          "Data"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamSubscriptionRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamSubscriptionRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StreamSubscriptionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{sessionId}/unsubscribe": {
      "post": {
        "tags": [
          "Data"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamSubscriptionRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamSubscriptionRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StreamSubscriptionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/sessions/{id}": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sessions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AggregationRequestDto": {
        "required": [
          "queries"
        ],
        "type": "object",
        "properties": {
          "queries": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/AggregationQueryDto"
            }
          }
        }
      },
      "AggregationResponseDto": {
        "required": [
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/AggregationResultDto"
            }
          }
        }
      },
...
      "ArchetypeDto": {
        "required": [
          "archetypeId",
          "name"
        ],
        "type": "object",
        "properties": {
          "archetypeId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint16"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ComponentTypeDto": {
        "required": [
          "componentTypeId",
          "name"
        ],
        "type": "object",
        "properties": {
          "componentTypeId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StreamSubscriptionRequestDto": {
        "required": [
          "streamId",
          "events"
        ],
        "type": "object",
        "properties": {
          "streamId": {
            "type": "string",
            "format": "uuid"
          },
          "events": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SystemListDto": {
        "required": [
          "query",
          "systems"
        ],
        "type": "object",
        "properties": {
          "query": {
            "type": [
              "null",
              "string"
            ]
          },
          "systems": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/SystemDefinitionDto"
            }
          }
        }
      },
...
      "TopologyDto": {
        "required": [
          "systems",
          "archetypes",
          "componentTypes",
          "phases"
        ],
        "type": "object",
        "properties": {
          "systems": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/SystemDefinitionDto"
            }
          },
          "archetypes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ArchetypeDto"
            }
          },
          "componentTypes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ComponentTypeDto"
            }
          },
          "phases": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      },
      "TrackDataResponseDto": {
        "required": [
          "trackId",
          "records"
        ],
        "type": "object",
        "properties": {
          "trackId": {
            "type": [
              "null",
              "string"
            ]
          },
          "records": {
            "type": [
              "null",
              "array"
            ]
          }
        }
      },
...
      "TracksResponseDto": {
        "required": [
          "tracks"
        ],
        "type": "object",
        "properties": {
          "tracks": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/TrackSchemaDto"
            }
          }
        }
      },
...
    }
  },
...
}